Functional tradebook!

+++ trading/risk.py
@@ -1,7 +1,11 @@
 #!/usr/bin/python3
 
+# This script calculates the risk, max position size, max share size and more and inputs 
+# them in the current existing TradeBook. If there is no tradebook the script will make one.
+
 import csv
 import pandas as pd
+from datetime import datetime
 
 # risk = float(input("\tRisk percentage:\t"))
 risk = 1
@@ -17,29 +21,68 @@ def truncate(n, decimals=0):
     return int(n * multiplier) / multiplier
 
 def TradeBook(tot_acc, enter, sl, risk, trade_type, pos, max_shares, sl_diff_p, max_loss):
+	# Function tries to first open the existing tradebook and otherwise create one. 
+	# Then it inserts the new trade depending on the answer of the user.
 	to_log = input("Do you want to log this trade? (y/n) ").lower()
-	try:
-		df = pd.read_csv("tradebook.csv")
-		print("Im using existing tradebook")
-	except:
-		df = pd.DataFrame({
-			'TradeNr':[], 
-			'Date':[], 
-			'AccBalance':[], 
-			'Enter':[],
-			'Stop Loss':[], 
-			'Risk':[], 
-			'Trade Type':[], 
-			'Position Size':[],
-			'Max_Shares':[],
-			'Sl_diff_p':[], 
-			'Max_Loss':[], 
-		})
-		df.to_csv(path_or_buf = "/mnt/c/Users/ThijmenWijgers/Documents/scripts/trading/tradebook.csv", index=False)
-
+	now = datetime.now()
+	now_formatted = now.strftime("%d-%m-%Y %H:%M:%S")
+	if to_log == 'y':
+		try:
+			df = pd.read_csv("tradebook.csv")
+			current_max_trade = df.max()["TradeNr"]
+			new_trade = {
+				'TradeNr':current_max_trade + 1, 
+				'Date':now_formatted, 
+				'Account Balance':tot_acc, 
+				'Enter':enter,
+				'Stop Loss':sl, 
+				'Total Acc Risk':f"{risk}%", 
+				'Trade Type':trade_type, 
+				'Position Size':truncate(pos),
+				'Max Shares':truncate(max_shares),
+				'Stop Loss %':f"{round(sl_diff_p*100, 1)}%", 
+				'Max. Loss':max_loss, 
+			}
+			df = df.append(new_trade, ignore_index=True)
+			df.to_csv(path_or_buf = "/mnt/c/Users/ThijmenWijgers/Documents/scripts/trading/tradebook.csv", index=False)
+			print(f"Your trade has been registered in existing tradebook as trade {current_max_trade +1}")
+		except:
+			df = pd.DataFrame({
+				'TradeNr':[], 
+				'Date':[], 
+				'Account Balance':[], 
+				'Total Acc Risk':[], 
+				'Enter':[],
+				'Stop Loss':[], 
+				'Stop Loss %':[], 
+				'Trade Type':[], 
+				'Position Size':[],
+				'Max Shares':[],
+				'Max. Loss':[], 
+			})
+			new_trade = {
+				'TradeNr':1, 
+				'Date':now_formatted, 
+				'Account Balance':tot_acc, 
+				'Enter':enter,
+				'Stop Loss':sl, 
+				'Total Acc Risk':f"{risk}%", 
+				'Trade Type':trade_type, 
+				'Position Size':truncate(pos),
+				'Max Shares':truncate(max_shares),
+				'Stop Loss %':f"{round(sl_diff_p*100, 1)}%", 
+				'Max. Loss':max_loss, 
+			}
+			df = df.append(new_trade, ignore_index=True)
+			overwrite = input("Are you sure you want to overwrite existing tradebook? (y/n) ").lower()
+			if overwrite == 'y':
+				df.to_csv(path_or_buf = "/mnt/c/Users/ThijmenWijgers/Documents/scripts/trading/tradebook.csv", index=False)
+				print("Creating new Tradebook and adding your first trade!")
 
 
 def calculateRisk(tot_acc, enter, sl, risk):
+	# Function initiates all the variables needed to calculate the risk and max position
+	# After calculation it inputs all the values in the TradeBook function
 	if enter > sl:
 		trade_type = "Long"
 	else:
@@ -58,7 +101,7 @@ def calculateRisk(tot_acc, enter, sl, risk):
 	print(f"\tMax Shares:\t{truncate(max_shares)}")
 	print(f"\tRisk Perc:\t{round(sl_diff_p*100, 2)}%")
 	print(f"\tMax Loss\t{max_loss}\n")
-	TradeBook()
+	TradeBook(tot_acc, enter, sl, risk, trade_type, pos, max_shares, sl_diff_p, max_loss)
 		
 if __name__ == "__main__":
 	calculateRisk(tot_acc, enter, sl, risk)
